Moved templated list functions
There was a weird workaround when importing compound tag.
