+++ src/amulet_nbt/include/amulet_nbt/tag/compound.hpp
@@ -10,6 +10,7 @@
 #include <amulet_nbt/tag/int.hpp>
 #include <amulet_nbt/tag/float.hpp>
 #include <amulet_nbt/tag/string.hpp>
+#include <amulet_nbt/tag/list.hpp>
 #include <amulet_nbt/tag/array.hpp>
 
 namespace AmuletNBT {
@@ -45,5 +46,3 @@ namespace AmuletNBT {
     template<> struct tag_id<CompoundTag> { static constexpr std::uint8_t value = 10; };
     template<> struct tag_id<CompoundTagPtr> { static constexpr std::uint8_t value = 10; };
 }
-
-#include <amulet_nbt/tag/list.hpp>
